# api/openapi.yaml
openapi: 3.1.0
info:
  title: ToolHub Browser API
  version: 2.0.0
  description: |
    Generation-2 single-user control-plane API. Unsafe authenticated requests
    require the session CSRF token. Operations that declare Idempotency-Key
    reject key reuse with a different request.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /nodes/refresh:
    post:
      tags: [Targets]
      operationId: refreshNodes
      parameters:
        - {$ref: '#/components/parameters/CSRFToken'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      responses:
        '202': {$ref: '#/components/responses/Operation'}
...
  /operations/{operationID}:
    get:
      tags: [Operations]
      operationId: getOperation
      parameters: [{$ref: '#/components/parameters/OperationID'}]
      responses:
        '200': {$ref: '#/components/responses/Operation'}
        '404': {$ref: '#/components/responses/Error'}
  /operations/{operationID}/cancel:
    post:
      tags: [Operations]
      operationId: cancelOperation
      parameters:
        - {$ref: '#/components/parameters/OperationID'}
        - {$ref: '#/components/parameters/CSRFToken'}
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '409': {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    sessionCookie: {type: apiKey, in: cookie, name: toolhub_session}
  parameters:
    CSRFToken: {name: X-CSRF-Token, in: header, required: true, schema: {type: string}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 8, maxLength: 200}}
...
    OperationID: {name: operationID, in: path, required: true, schema: {$ref: '#/components/schemas/UUID'}}
...
  responses:
    NoContent: {description: No content}
    Error:
      description: Error envelope
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
...
    Operation: {description: Durable operation, content: {application/json: {schema: {$ref: '#/components/schemas/Operation'}}}}
...
  schemas:
    UUID: {type: string, format: uuid}
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: {type: string}
            message: {type: string}
            requestId: {type: string}
            details: {type: object}
...
    Operation:
      type: object
      required: [id, kind, status, metadata, cancelRequested, createdAt, updatedAt]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        kind: {type: string, enum: [skill_import, mcp_import, update_check, apply, edit, restore, scan, refresh, reconcile, relay_start, relay_stop, relay_restart, relay_health, backup_gc]}
        status: {$ref: '#/components/schemas/OperationStatus'}
        sourceId: {$ref: '#/components/schemas/UUID'}
        metadata: {type: object}
        errorCode: {type: string}
        errorReason: {type: string}
        cancelRequested: {type: boolean}
        targets: {type: array, items: {$ref: '#/components/schemas/OperationTarget'}}
        createdAt: {type: string, format: date-time}
        startedAt: {type: [string, 'null'], format: date-time}
        finishedAt: {type: [string, 'null'], format: date-time}
        updatedAt: {type: string, format: date-time}
    OperationTarget:
      type: object
      required: [id, targetId, targetKey, status, attempt, pendingRerun]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        targetId: {$ref: '#/components/schemas/UUID'}
        targetKey: {type: string}
        status: {type: string, enum: [queued, running, succeeded, failed, cancelled]}
        attempt: {type: integer, minimum: 1}
        pendingRerun: {type: boolean}
        bridgeOperationId: {$ref: '#/components/schemas/UUID'}
        saltJid: {type: string, pattern: '^[0-9]{8,32}$'}
        result: {type: object, description: Redacted safe terminal result returned by the typed Bridge operation.}
        errorCode: {type: string}
        errorReason: {type: string}
        createdAt: {type: string, format: date-time}
        startedAt: {type: [string, 'null'], format: date-time}
        finishedAt: {type: [string, 'null'], format: date-time}
        updatedAt: {type: string, format: date-time}
...
    OperationStatus: {type: string, enum: [queued, running, succeeded, partial, failed, cancelled]}